requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
firebase-admin==6.5.0
groq==0.9.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.7.4
numpy==1.26.4
torch==2.3.0
transformers==4.41.0